+++ packages/outbox/src/dispatchers/event-queue.dispatcher.ts
@@ -1,0 +1,7 @@
+import {
+  type EventQueueEntity,
+  type EventQueueModel,
+  OutboxDispatcher,
+} from '@volontariapp/database';
+
+export class EventQueueDispatcher extends OutboxDispatcher<EventQueueModel, EventQueueEntity> {}